+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## [Unreleased](https://github.com/Piturnah/gex/compare/v0.3.7...master)
+### Changed
+- Cursor switches to bar when navigating left and right in git command
+
 ## [0.3.7](https://github.com/Piturnah/gex/compare/v0.3.6...v0.3.7) - 2023-01-25
 ### Added
 - Improvements to arbitary git command execution